renamed files and endpoint
